bcm4354-bt: Adapt to Resin platform
The previous meta-artik bsp update brought in bsp support.
However, the standard Samsung way is storing some configuration files in /opt. For
using it with our resinhup tool, we need this changed in a location that will not
be read-only and we can also instruct resinhup to ignore the files in this location
(the resin-conf partition to be precise) so that the bluetooth tools can write to
those files if necessary.

This commit patches the pristine Samsung bluetooth configuration files/scripts so
that /mnt/conf/opt directory is used instead of the /opt directory.

Signed-off-by: Florin Sarbu <florin@resin.io>