kernel-modules-headers.bbappend: Add patch for removing the option to have the script output in a different directory.
Because we cannot use "make O=" with an un-clean kernel source tree
(such as for the artik boards) we need to patch the kernel modules header
generation script to remove this option from make.
At the same time remove unnecessary copying of .config and Module.symvers
to target directory since that is only needed when using "make O=" to target
directory and we do not use that here.

Signed-off-by: Florin Sarbu <florin@resin.io>