balena-image.inc: Set emgw3 rootfs size to 400 MiB
The honister update increased the rootfs size for
nitrogen 8MM based devices and we had to resort to
firmware compression to work around the issue.
In order to avoid such issues for the new emgw3 board,
let's start from scratch with a 400 MiB rootfs size
instead of the 320 MiB default one.

Signed-off-by: Florin Sarbu <florin@balena.io>