u-boot: Fixes for kirkstone update
With the move to kirkstone we need further fixes related to upstream BSP
u-boot config changes. We disable CONFIG_ENV_IS_IN_MMC as we do not use it
and it also throws a compile error.
Additionaly, we remove the disabling of CONFIG_CMD_SAVEENV from here as it
is already disabled globally from meta-balena.
And last, we update the bootscript by increasing the fdt load address in order
to account for increased kernel size.

Signed-off-by: Florin Sarbu <florin@balena.io>