Revert "layers/meta-balena-fsl-arm/conf/layer.conf: Compress kernel for nitrogen8mm"
This reverts commit b8cc3390f3bd3c763d5b70f96a663b5cfdf39898.

We now have enabled kernel module compression in meta-balena which saves us enough
space so we can revert the kernel image compression. The kernel image compression
commit also needed additional changes in u-boot which were not made so for now it is
faster to fix it by reverting the kernel compression.

Changelog-entry: Revert kernel compression for nitrogen8mm based boards
Signed-off-by: Florin Sarbu <florin@balena.io>