hostapp-update-hooks: Replace u-boot with newer one when updating from v2.114.0
The rev3.0 Nitrogen8MM DWE has no problems rebooting with older
OSes like v2.114.0 or older. However, the rev2.0 board in the
autokit has this problem, which can be fixed by updating u-boot.

The new u-boot does not exhibit the reboot problem, neither with
the rev2.0 nor with 3.0.

This workaround helps running rollback and HUP tests in autokit,
where a rev2.0 board is used.

Signed-off-by: Alexandru Costache <alexandru@balena.io>