emgw3.coffee: Specify the flasher artifact
For automated testing we need to use the flasher image to provision the board.
The customer will keep using the internal image for provisioning, which is
the image that will continue to be offered in the dashboard.

Changelog-entry: Specify in the coffee file what is the flasher image for emgw3
Signed-off-by: Florin Sarbu <florin@balena.io>